+++ tests/testthat/test-write-docker.R
@@ -3,7 +3,7 @@ b <- board_folder(path = tmp_dir)
 
 test_that("create Dockerfile with packages", {
     skip_on_cran()
-    v$metadata$required_pkgs <- c("beepr", "caret", "stats")
+    v$metadata$required_pkgs <- c("pingr", "caret", "stats")
     vetiver_pin_write(b, v)
     vetiver_write_plumber(b, "cars1", file = file.path(tmp_dir, "plumber.R"))
     vetiver_write_docker(v, file.path(tmp_dir, "plumber.R"), tmp_dir)
@@ -13,9 +13,22 @@ test_that("create Dockerfile with packages", {
     )
 })
 
+test_that("create Dockerfile with 'additional' packages", {
+    skip_on_cran()
+    v$metadata$required_pkgs <- c("pingr", "caret", "stats")
+    vetiver_pin_write(b, v)
+    vetiver_write_plumber(b, "cars1", file = file.path(tmp_dir, "plumber.R"))
+    vetiver_write_docker(v, file.path(tmp_dir, "plumber.R"), tmp_dir,
+                         additional_pkgs = c("caret", "ggplot2"))
+    expect_snapshot(
+        cat(readr::read_lines(file.path(tmp_dir, "Dockerfile")), sep = "\n"),
+        transform = redact_vetiver
+    )
+})
+
 test_that("create Dockerfile with no RSPM", {
     skip_on_cran()
-    v$metadata$required_pkgs <- c("beepr", "caret")
+    v$metadata$required_pkgs <- c("pingr", "caret")
     vetiver_pin_write(b, v)
     vetiver_write_plumber(b, "cars1", file = file.path(tmp_dir, "plumber.R"))
     vetiver_write_docker(v, file.path(tmp_dir, "plumber.R"),
@@ -39,7 +52,7 @@ test_that("create Dockerfile with no packages", {
 
 test_that("create Dockerfile with specific port", {
     skip_on_cran()
-    v$metadata$required_pkgs <- c("beepr", "caret")
+    v$metadata$required_pkgs <- c("pingr", "caret")
     vetiver_pin_write(b, v)
     vetiver_write_plumber(b, "cars1", file = file.path(tmp_dir, "plumber.R"))
     vetiver_write_docker(v, file.path(tmp_dir, "plumber.R"),